+++ release_notes.md
@@ -14,3 +14,4 @@
 - Avoid setting up OTel and AzMon exporter in the placeholder mode. (#11090)
 - Update Java Worker Version to [2.19.1](https://github.com/Azure/azure-functions-java-worker/releases/tag/2.19.1)
 - Memory allocation optimizations in `FeatureFlags.IsEnabled` by adopting zero-allocation `ContainsToken` for efficient delimited token search (#11075)
+- Improved metadata binding validation (#11101)